requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.0.0
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.0
psycopg2-binary==2.8.5
pytz==2020.1
redis==3.5.2
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
gunicorn==20.0.4
requests==2.23.0
# Documentation
sphinx
sphinx-copybutton